# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel HRM Documentation",
        "description": "L5 Swagger OpenApi description for hrm project",
        "contact": {
            "email": "drrifatalashwad0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Student John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.student@yopmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678@We"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "student"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "afd542192a0c3f60c1af223b81e87f89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.student@yopmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678@We"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "courses/{id}/complete": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Generate a certificate for a completed course",
                "operationId": "e5e28a14c5ddb57b4787cd2fb54a9a60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 101
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Certificate generated successfully"
                    },
                    "400": {
                        "description": "Course not completed yet"
                    }
                }
            }
        },
        "certificates/download/{id}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Download a certificate PDF",
                "operationId": "a52f3cb7efc12fe8db1486fab4905e09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Certificate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate PDF download"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            }
        },
        "certificates/verify/{code}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Verify a certificate by code",
                "operationId": "5bba722af19943ac54bc065ddebcb257",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Certificate code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ABCD123XYZ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate is valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "course": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Laravel Basics"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "issued_at": {
                                            "type": "string",
                                            "example": "2025-09-16 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Certificate not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}